+++ src/lib/create_waku.ts
@@ -6,9 +6,10 @@ import { all as filterAll } from "@libp2p/websockets/filters";
 import { createLibp2p, Libp2pOptions } from "libp2p";
 import type { Libp2p } from "libp2p";
 
+import type { Waku } from "./interfaces";
 import { PeerDiscoveryStaticPeers } from "./peer_discovery_static_list";
 import { getPredefinedBootstrapNodes } from "./predefined_bootstrap_nodes";
-import { Waku, WakuOptions } from "./waku";
+import { WakuNode, WakuOptions } from "./waku";
 import { WakuFilter } from "./waku_filter";
 import { WakuLightPush } from "./waku_light_push";
 import { WakuRelay } from "./waku_relay";
@@ -30,11 +31,11 @@ export interface CreateOptions {
    */
   pubSubTopic?: string;
   /**
-   * You can pass options to the `Libp2p` instance used by {@link index.waku.Waku} using the {@link CreateOptions.libp2p} property.
+   * You can pass options to the `Libp2p` instance used by {@link index.waku.WakuNode} using the {@link CreateOptions.libp2p} property.
    * This property is the same type than the one passed to [`Libp2p.create`](https://github.com/libp2p/js-libp2p/blob/master/doc/API.md#create)
    * apart that we made the `modules` property optional and partial,
    * allowing its omission and letting Waku set good defaults.
-   * Notes that some values are overridden by {@link index.waku.Waku} to ensure it implements the Waku protocol.
+   * Notes that some values are overridden by {@link index.waku.WakuNode} to ensure it implements the Waku protocol.
    */
   libp2p?: Partial<Libp2pOptions>;
   /**
@@ -65,7 +66,13 @@ export async function createWaku(
   const wakuLightPush = new WakuLightPush(libp2p, options);
   const wakuFilter = new WakuFilter(libp2p, options);
 
-  return new Waku(options ?? {}, libp2p, wakuStore, wakuLightPush, wakuFilter);
+  return new WakuNode(
+    options ?? {},
+    libp2p,
+    wakuStore,
+    wakuLightPush,
+    wakuFilter
+  );
 }
 
 export function defaultPeerDiscovery(): PeerDiscovery {